moved AsP statbox to AstralPowerDetails and updated about tab
